Fix sending images from bots
It appears that GroupMe may have changed their API for bots since this
was first written. Rather than taking an options map with a picture_url,
the /bots/post method takes a picture_url directly for posting images
with bots. So, fix the implementation here to make it match what GroupMe
expects.