Fix exception raised when liking messages
When liking and unliking messages, the body of the GroupMe response
contains a 'meta' field with a status code and a 'response' field that
evaluates to nil. Because the body was non-empty, we would return a nil
response to users of the 'request' method when they expected the full
response from Faraday. In particular, this led to an exception being
thrown when liking messages (although messages would still be liked
successfully).

This fix should allow create_like and destroy_like to work properly.